syntax: string: Do not treat bare "[" as special.
Fish only recognizes square brackets as special when they follow a
variable name, as in "$array[1]".  By default, contained matches and
regions extend the parent region.  As `fishDerefExtension` requires a
closing "]", a string with a bare "[" would not be recoginzed as a
terminated one.

Simples example would be a string like this: "a[b".

As square brackets have no special meaning, unless they follow a
variable name, there seems to be no reason for highlighting them.